refactor(web-app): create TimeRangeDropdown component
