+++ packages/cube-frontend-web-app/src/components/TimeRangeDropdown/timeRangeDropdownUtils.ts
@@ -1,0 +1,9 @@
+export const timeRanges = [
+  'last30Days',
+  'last14Days',
+  'last7Days',
+  'last24Hours',
+  'lastHour',
+] as const
+
+export type TimeRange = (typeof timeRanges)[number]
