+++ packages/cube-frontend-web-app/src/pages/home/health/_components/healthAccordion/HealthAccordion.tsx
@@ -1,23 +1,13 @@
-import { CosDropdown } from '@cube-frontend/ui-library'
+import { TimeRangeDropdown } from '@cube-frontend/web-app/components/TimeRangeDropdown/TimeRangeDropdown'
+import { TimeRange } from '@cube-frontend/web-app/components/TimeRangeDropdown/timeRangeDropdownUtils'
 import { useMemo, useState } from 'react'
 import { HealthAccordionItem } from './HealthAccordionItem'
 import {
   groupServicesByCategory,
   ServiceCategory,
-  TimeRange,
-  timeRanges,
 } from './healthAccordionUtils'
 import { mockServices } from './mockHealth'
 import { useNow } from './useNow'
-
-// TODO: Replace this with i18n.
-const timeRangeLabels: Record<TimeRange, string> = {
-  last30Days: 'Last 30 days',
-  last14Days: 'Last 14 days',
-  last7Days: 'Last 7 days',
-  last24Hours: 'Last 24 hours',
-  lastHour: 'Last hour',
-}
 
 export const HealthAccordion = () => {
   const [selectedTimeRange, setSelectedTimeRange] =
@@ -49,22 +39,10 @@ export const HealthAccordion = () => {
         <span className="secondary-body1 text-functional-text-light">
           Details
         </span>
-        <CosDropdown selectedItems={[selectedTimeRange]}>
-          <CosDropdown.Trigger>
-            {timeRangeLabels[selectedTimeRange]}
-          </CosDropdown.Trigger>
-          <CosDropdown.Menu>
-            {timeRanges.map((timeRange) => (
-              <CosDropdown.Item
-                key={timeRange}
-                item={timeRange}
-                onClick={() => setSelectedTimeRange(timeRange)}
-              >
-                {timeRangeLabels[timeRange]}
-              </CosDropdown.Item>
-            ))}
-          </CosDropdown.Menu>
-        </CosDropdown>
+        <TimeRangeDropdown
+          selectedItem={selectedTimeRange}
+          onChange={setSelectedTimeRange}
+        />
       </div>
       {categories.map((category) => (
         <HealthAccordionItem
