+++ packages/cube-frontend-web-app/src/pages/home/health/_components/healthAccordion/healthAccordionUtils.ts
@@ -25,13 +25,3 @@ export const groupServicesByCategory = (
 
   return Array.from(map.values())
 }
-
-export const timeRanges = [
-  'last30Days',
-  'last14Days',
-  'last7Days',
-  'last24Hours',
-  'lastHour',
-] as const
-
-export type TimeRange = (typeof timeRanges)[number]
